Adopt changes to Read and Diff
1. Return resource inputs as well as resource state from Read()
2. Return a list of properties that changed from Diff()

We implement the former by copying all properties from a resource's
state to the input bag and removing the following well-known read-only
fields:

- metadata.{uid,resourceVersion,generation,selfLink,continue,
  creationTimestamp,deletionTimestamp,deletionGracePeriodSeconds}
- status

The result is a set of inputs that is sufficient (if perhaps not
necessary) to create a resource with a state that is identical to the
resource's current state. This is used to detect drift during `refresh`
and to implement resource import.

We implement the latter by scanning the JSON diff and recording the
names of the top-level properties that changed. The engine uses this
information to filter diffs to only those that are semantically
meaningful.

These changes required a couple of bugfixes:
- Old names are only adopted if the old resource was auto-named. This
  ensures that a name must be specified when importing a resource that
  was not autonamed.
- URN to GVK conversion was fixed for resources in the "core" group.
  These resources have no group part in the GVK. Parsing was also
  simplified through the use of pulumi/pulumi's token manipulation
  functions.
- When reading a resource, the GVK for the resource to read is now
  pulled from the URN if it is absent from the inputs.